+++ dhcpv4/bindtointerface.go
@@ -1,0 +1,10 @@
+package dhcpv4
+
+import (
+	"github.com/insomniacslk/dhcp/interfaces"
+)
+
+// BindToInterface (deprecated) redirects to interfaces.BindToInterface
+func BindToInterface(fd int, ifname string) error {
+	return interfaces.BindToInterface(fd, ifname)
+}
